client: move all video players into their own folder
